requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
python-dotenv==0.19.0
pandas==1.3.3
plotly==5.3.1
python-binance==1.0.15
requests==2.26.0 